Fixes #71: Take server path prefix into consideration.
- Refactored network router code.
- NoteRouter.swift got split up into its individual types.
- OCSRouter and StatusRouter now also take the server path prefix into consideration.
- Removed dead code in NotesTableViewController due to new login introduced recently.
