+++ iOCNotes.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "1658e0ccffb421f61f86533c4021dfa389b4e4f3552ab84e4ae1441892b317e0",
+  "originHash" : "f47d5350204810d2423a1d61e8832fd4e9af60c6d726f7efd130b4b279958a8d",
   "pins" : [
     {
       "identity" : "alamofire",
